+++ scripts/configs/types.ts
@@ -133,6 +133,8 @@ export interface NetworkAddresses<T extends Record<string, AddressInfo> = {}> {
     PROTOCOL_GUARDIAN?: Hex;
     AAVE_MERKLE_DISTRIBUTOR?: Hex;
     AAVE_CL_ROBOT_OPERATOR?: Hex;
+    ROBOT_GUARDIAN?: Hex;
+    AFC_SAFE?: Hex;
     AGENT_HUB?: Hex;
     AGENT_HUB_AUTOMATION?: Hex;
     RANGE_VALIDATION_MODULE?: Hex;
@@ -152,6 +154,7 @@ export interface V4Config {
   admin: Record<string, Hex>;
   hubs: Record<string, Hex>;
   spokes: Record<string, Hex>;
+  eSpokes?: Record<string, Hex>;
   positionManagers?: Record<string, Hex>;
   externalLibraries: Record<string, Hex>;
   additionalAddresses?: Record<string, Hex>;