+++ tests/sanity/psp.spec.ts
@@ -1,8 +1,8 @@
 import {describe, it, expect} from 'vitest';
-import * as addressBook from '../../src/ts/AaveAddressBook';
+import * as addressBook from 'src/ts/AaveAddressBook';
 import {getContract} from 'viem';
-import {getClient} from '../../scripts/clients';
-import {getGovernance, getWhiteLabelGovernance, isPoolWhiteLabel} from '../utils';
+import {getClient} from 'scripts/clients';
+import {getGovernance, getWhiteLabelGovernance, isPoolWhiteLabel} from 'tests/utils';
 
 export async function checkGetters(address, addresses: Record<string, any>) {
   const client = getClient(addresses.CHAIN_ID);