+++ CHANGELOG.md
@@ -2,6 +2,90 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [4.52.3](https://github.com/aave-dao/aave-address-book/compare/v4.52.2...v4.52.3) (2026-05-22)
+
+
+### Bug Fixes
+
+* **cache:** automated cache update - updated addresses ([#1398](https://github.com/aave-dao/aave-address-book/issues/1398)) ([1972360](https://github.com/aave-dao/aave-address-book/commit/19723607ecbf352ab374590c8b993f3e3dbca744))
+
+## [4.52.2](https://github.com/aave-dao/aave-address-book/compare/v4.52.1...v4.52.2) (2026-05-21)
+
+
+### Bug Fixes
+
+* add better config for treeshaking with modern bundlers ([#1394](https://github.com/aave-dao/aave-address-book/issues/1394)) ([0a61c1f](https://github.com/aave-dao/aave-address-book/commit/0a61c1f17d7b8d08d859be788ff0152a70edd43f))
+
+## [4.52.1](https://github.com/aave-dao/aave-address-book/compare/v4.52.0...v4.52.1) (2026-05-20)
+
+
+### Bug Fixes
+
+* **deps:** bump lib/aave-v3-origin from `1e3d70c` to `f6f9cfc` ([#1395](https://github.com/aave-dao/aave-address-book/issues/1395)) ([0181771](https://github.com/aave-dao/aave-address-book/commit/0181771ce9e3ad0422e11dbf81e684345d6b3daf))
+
+## [4.52.0](https://github.com/aave-dao/aave-address-book/compare/v4.51.0...v4.52.0) (2026-05-19)
+
+
+### Features
+
+* allSpokesRaw getter for v4 ([#1392](https://github.com/aave-dao/aave-address-book/issues/1392)) ([ac6a932](https://github.com/aave-dao/aave-address-book/commit/ac6a9326bdd7d2e19240bf574fb5c758a37833eb))
+
+## [4.51.0](https://github.com/aave-dao/aave-address-book/compare/v4.50.0...v4.51.0) (2026-05-18)
+
+
+### Features
+
+* v4 getter library ([#1383](https://github.com/aave-dao/aave-address-book/issues/1383)) ([fe21ba4](https://github.com/aave-dao/aave-address-book/commit/fe21ba4e186a5982a24d791f55c9fd78fded440e))
+
+## [4.50.0](https://github.com/aave-dao/aave-address-book/compare/v4.49.13...v4.50.0) (2026-05-18)
+
+
+### Features
+
+* Add sGho deployment ([#1377](https://github.com/aave-dao/aave-address-book/issues/1377)) ([6eac158](https://github.com/aave-dao/aave-address-book/commit/6eac158a825aaf49126f9cb036024183b400f190))
+
+## [4.49.13](https://github.com/aave-dao/aave-address-book/compare/v4.49.12...v4.49.13) (2026-05-17)
+
+
+### Bug Fixes
+
+* **cache:** automated cache update - updated addresses ([#1386](https://github.com/aave-dao/aave-address-book/issues/1386)) ([44303bb](https://github.com/aave-dao/aave-address-book/commit/44303bb4d2b61c12af2cb7bde7eb6e7aa56fdbcd))
+
+## [4.49.12](https://github.com/aave-dao/aave-address-book/compare/v4.49.11...v4.49.12) (2026-05-16)
+
+
+### Bug Fixes
+
+* **cache:** automated cache update - updated addresses ([#1384](https://github.com/aave-dao/aave-address-book/issues/1384)) ([b9cd3b7](https://github.com/aave-dao/aave-address-book/commit/b9cd3b7287852e3a9ff488df5b216e15c25c357b))
+
+## [4.49.11](https://github.com/aave-dao/aave-address-book/compare/v4.49.10...v4.49.11) (2026-05-15)
+
+
+### Bug Fixes
+
+* **cache:** automated cache update - updated addresses ([#1381](https://github.com/aave-dao/aave-address-book/issues/1381)) ([1f88ec1](https://github.com/aave-dao/aave-address-book/commit/1f88ec131bdea1fc24cc272f16b8a490f0443935))
+
+## [4.49.10](https://github.com/aave-dao/aave-address-book/compare/v4.49.9...v4.49.10) (2026-05-10)
+
+
+### Bug Fixes
+
+* **cache:** automated cache update - updated addresses ([#1379](https://github.com/aave-dao/aave-address-book/issues/1379)) ([06dd579](https://github.com/aave-dao/aave-address-book/commit/06dd57954698c6325affa8e4d77ab3ea0c0e5684))
+
+## [4.49.9](https://github.com/aave-dao/aave-address-book/compare/v4.49.8...v4.49.9) (2026-05-07)
+
+
+### Bug Fixes
+
+* rename eSpokes & small fixes ([#1368](https://github.com/aave-dao/aave-address-book/issues/1368)) ([4ee7b50](https://github.com/aave-dao/aave-address-book/commit/4ee7b505d40760f48bcfdc4edd5fe3d532ec88d2))
+
+## [4.49.8](https://github.com/aave-dao/aave-address-book/compare/v4.49.7...v4.49.8) (2026-05-01)
+
+
+### Bug Fixes
+
+* **cache:** automated cache update - updated addresses ([#1375](https://github.com/aave-dao/aave-address-book/issues/1375)) ([2b5f43f](https://github.com/aave-dao/aave-address-book/commit/2b5f43fb2543790cbe8faf459f4b53a9cca0334a))
+
 ## [4.49.7](https://github.com/aave-dao/aave-address-book/compare/v4.49.6...v4.49.7) (2026-04-25)